GetFromType and related moved to different helper in ExportModule.Data
