+++ Modules/vc-module-export/VirtoCommerce.ExportModule.Data/Extensions/ExportDataQueryExtensions.cs
@@ -17,8 +17,8 @@ namespace VirtoCommerce.ExportModule.Data.Extensions
             {
                 foreach (var property in type.GetProperties().Where(x => x.CanRead && x.CanWrite))
                 {
-                    var propertyName = ExportedTypeMetadata.GetDerivedName(baseMemberName, property);
-                    var nestedType = ExportedTypeMetadata.GetNestedType(property.PropertyType);
+                    var propertyName = property.GetDerivedName(baseMemberName);
+                    var nestedType = property.PropertyType.GetNestedType();
 
                     if (nestedType.IsSubclassOf(typeof(Entity)))
                     {
