+++ Modules/vc-module-export/VirtoCommerce.ExportModule.Tests/CsvExportProviderTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Text;
@@ -383,7 +384,7 @@ namespace VirtoCommerce.ExportModule.Tests
                 }
             };
 
-            var configuration = new Configuration();
+            var configuration = new Configuration(cultureInfo: CultureInfo.GetCultureInfo("en-US"));
             configuration.RegisterClassMap<PricelistTestMapping>();
 
             //Act
@@ -459,7 +460,7 @@ namespace VirtoCommerce.ExportModule.Tests
 
         private string SerializeAndRead(ExportedTypeMetadata metadata, IEnumerable items, Configuration configuration = null)
         {
-            var csvConfiguration = new CsvProviderConfiguration() { Configuration = configuration ?? new Configuration()};
+            var csvConfiguration = new CsvProviderConfiguration() { Configuration = configuration ?? new Configuration(cultureInfo: CultureInfo.GetCultureInfo("en-US")) };
 
             using (var stream = new MemoryStream())
             using (var writer = new StreamWriter(stream, Encoding.UTF8) { AutoFlush = true })
